refactor: change protocol to more generic

+++ src/data/usecases/sector/db-delete-sector.ts
@@ -1,19 +1,20 @@
-import { DeleteSectorRepository, CheckOwnerBySectorIdRepository } from '@/data/protocols'
+import { DeleteSectorRepository, CheckDataByFieldRepository } from '@/data/protocols'
 import { DeleteSector } from '@/domain/usecases'
 
 export class DbDeleteSector implements DeleteSector {
   constructor (
     private readonly deleteSectorRepository: DeleteSectorRepository,
-    private readonly checkOwnerBySectorIdIdRepository: CheckOwnerBySectorIdRepository
+    private readonly checkOwnerBySectorIdIdRepository: CheckDataByFieldRepository
   ) {}
 
   async delete (params: DeleteSector.Params): Promise<DeleteSector.Model> {
-    const { id } = params
-    const exists = await this.checkOwnerBySectorIdIdRepository.checkBySectorId({
-      sectorId: id
+    const { id, accountId } = params
+    const exists = await this.checkOwnerBySectorIdIdRepository.checkByField({
+      value: id,
+      accountId
     })
     if (!exists) {
-      return this.deleteSectorRepository.delete({ id })
+      return this.deleteSectorRepository.delete({ id, accountId })
     }
     return null
   }
